# sdks/sdkwork-web-backend-sdk/openapi/web-backend-api.openapi.json
{
  "openapi": "3.1.2",
  "info": {
    "title": "SDKWork Web Backend API",
    "description": "SDKWork Web Server management and control-plane API.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://api.web.sdkwork.com",
      "description": "Production"
    },
    {
      "url": "http://localhost:3800",
      "description": "Local Development"
    }
  ],
  "security": [
    {
      "AuthToken": [],
      "AccessToken": []
    }
  ],
  "paths": {
...
    "/backend/v3/api/nginx/etc/{configId}": {
      "parameters": [
        {
          "name": "configId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "operationId": "nginx.configs.retrieve",
        "summary": "Retrieve an Nginx configuration",
        "tags": [
          "nginx"
        ],
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-api-authority": "sdkwork-web.backend",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web.nginx.write",
        "x-sdkwork-owner": "sdkwork-web-server",
        "responses": {
          "200": {
            "description": "Nginx configuration returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NginxConfigResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-route-auth": "dual-token"
      },
      "put": {
        "operationId": "nginx.configs.update",
        "summary": "Update an Nginx configuration",
        "tags": [
          "nginx"
        ],
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-api-authority": "sdkwork-web.backend",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web.nginx.write",
        "x-sdkwork-idempotent": true,
        "x-sdkwork-owner": "sdkwork-web-server",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNginxConfigRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Nginx configuration updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NginxConfigResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Resource not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-route-auth": "dual-token"
      }
    },
...
    "/backend/v3/api/nginx/status": {
      "get": {
        "operationId": "nginx.status.retrieve",
        "summary": "Retrieve Nginx status",
        "tags": [
          "nginx"
        ],
        "security": [
          {
            "AuthToken": [],
            "AccessToken": []
          }
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-api-authority": "sdkwork-web.backend",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-auth-mode": "dual-token",
        "x-sdkwork-permission": "web.nginx.write",
        "x-sdkwork-owner": "sdkwork-web-server",
        "responses": {
          "200": {
            "description": "Nginx status returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/NginxStatusResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "x-sdkwork-route-auth": "dual-token"
      }
    },
...
    "/backend/v3/api/agent/heartbeat": {
      "post": {
        "operationId": "agent.heartbeat",
        "summary": "Report an edge-agent heartbeat",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "AgentToken": []
          }
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-api-authority": "sdkwork-web.backend",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-auth-mode": "api-key",
        "x-sdkwork-route-auth": "agent-token",
        "x-sdkwork-owner": "sdkwork-web-server",
        "x-sdkwork-rate-limit-tier": "auth-critical",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentHeartbeatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Edge-agent heartbeat acknowledged successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/AgentHeartbeatResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-permission": "web.agent.write"
      }
    },
    "/backend/v3/api/agent/sync": {
      "get": {
        "operationId": "agent.retrieve",
        "summary": "Retrieve the Nginx configuration and certificate bundle",
        "tags": [
          "agent"
        ],
        "security": [
          {
            "AgentToken": []
          }
        ],
        "x-sdkwork-api-surface": "backend-api",
        "x-sdkwork-api-authority": "sdkwork-web.backend",
        "x-sdkwork-request-context": "WebRequestContext",
        "x-sdkwork-auth-mode": "api-key",
        "x-sdkwork-route-auth": "agent-token",
        "x-sdkwork-owner": "sdkwork-web-server",
        "x-sdkwork-rate-limit-tier": "auth-critical",
        "parameters": [
          {
            "name": "ifSyncVersion",
            "in": "query",
            "description": "When it matches the current tenant manifest fingerprint, returns unchanged=true with empty bundles.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Edge-agent synchronization bundle returned successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SdkWorkApiResponse"
                    },
                    {
                      "type": "object",
                      "required": [
                        "data"
                      ],
                      "properties": {
                        "data": {
                          "type": "object",
                          "required": [
                            "item"
                          ],
                          "properties": {
                            "item": {
                              "$ref": "#/components/schemas/AgentSyncResponse"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        },
        "x-sdkwork-permission": "web.agent.read"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "AccessToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Access-Token"
      },
      "AgentToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-SDKWork-Agent-Token"
      }
    },
...
    "schemas": {
      "ProblemDetail": {
        "type": "object",
        "additionalProperties": true,
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "properties": {
          "type": {
            "type": "string",
            "format": "uri-reference"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer",
            "minimum": 100,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "code": {
            "$ref": "#/components/schemas/SdkWorkPlatformErrorCode"
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          }
        }
      },
...
      "UpdateNginxConfigRequest": {
        "type": "object",
        "properties": {
          "configContent": {
            "type": "string"
          },
          "configName": {
            "type": "string",
            "maxLength": 200
          }
        }
      },
      "NginxConfigResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "configType": {
            "type": "integer"
          },
          "configName": {
            "type": "string"
          },
          "configContent": {
            "type": "string"
          },
          "configHash": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "versionNo": {
            "type": "integer"
          },
          "deployedAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "NginxStatusResponse": {
        "type": "object",
        "properties": {
          "running": {
            "type": "boolean"
          },
          "version": {
            "type": "string"
          },
          "pid": {
            "type": "integer"
          },
          "activeConnections": {
            "type": "integer"
          },
          "configPath": {
            "type": "string"
          },
          "uptime": {
            "type": "string"
          }
        }
      },
...
      "AgentHeartbeatRequest": {
        "type": "object",
        "properties": {
          "agentVersion": {
            "type": "string",
            "maxLength": 100
          },
          "nginxEnabled": {
            "type": "boolean"
          },
          "activeConfigs": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "i64",
            "description": "Number of active nginx configs reported by the agent as a string."
          },
          "lastSyncVersion": {
            "type": "string",
            "pattern": "^sv1:[a-f0-9]{64}$",
            "description": "Last successfully applied syncVersion reported by the edge agent."
          }
        }
      },
      "AgentHeartbeatResponse": {
        "type": "object",
        "required": [
          "serverId",
          "status",
          "acknowledgedAt"
        ],
        "properties": {
          "serverId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "integer"
          },
          "acknowledgedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AgentSyncResponse": {
        "type": "object",
        "required": [
          "serverId",
          "syncVersion",
          "unchanged",
          "nginxConfigs",
          "certificates"
        ],
        "properties": {
          "serverId": {
            "type": "string",
            "format": "uuid"
          },
          "syncVersion": {
            "type": "string",
            "pattern": "^sv1:[a-f0-9]{64}$",
            "description": "Stable SHA-256 fingerprint of active nginx configs and certificates for the tenant."
          },
          "unchanged": {
            "type": "boolean",
            "description": "True when ifSyncVersion matched syncVersion; bundles are omitted to save bandwidth."
          },
          "nginxConfigs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentNginxConfigBundle"
            }
          },
          "certificates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AgentCertificateBundle"
            }
          }
        }
      },
      "AgentNginxConfigBundle": {
        "type": "object",
        "required": [
          "configId",
          "domain",
          "configContent",
          "fingerprint",
          "version"
        ],
        "properties": {
          "configId": {
            "type": "string",
            "format": "uuid"
          },
          "domain": {
            "type": "string",
            "minLength": 1,
            "maxLength": 253
          },
          "configContent": {
            "type": "string",
            "maxLength": 1048576
          },
          "fingerprint": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "SHA-256 hex digest of configContent."
          },
          "version": {
            "type": "string",
            "format": "int64",
            "pattern": "^[0-9]+$",
            "x-sdkwork-int64-string": true,
            "x-sdkwork-rust-type": "i64",
            "description": "Config revision number as a string to avoid JavaScript precision loss."
          }
        }
      },
      "AgentCertificateBundle": {
        "type": "object",
        "required": [
          "certificateId",
          "certName",
          "fingerprint",
          "fullchainPem",
          "privkeyPem"
        ],
        "properties": {
          "certificateId": {
            "type": "string",
            "format": "uuid"
          },
          "certName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 253
          },
          "fingerprint": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "fullchainPem": {
            "type": "string",
            "maxLength": 2097152
          },
          "privkeyPem": {
            "type": "string",
            "maxLength": 2097152
          }
        }
      },
...
      "SdkWorkApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "data",
          "traceId"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "enum": [
              0
            ],
            "default": 0,
            "minimum": 0,
            "maximum": 0
          },
          "data": {
            "description": "Operation-specific payload typed per response schema."
          },
          "traceId": {
            "type": "string",
            "format": "uuid",
            "description": "Server-owned request correlation id."
          }
        }
      },
...
      "SdkWorkPlatformErrorCode": {
        "type": "integer",
        "format": "int32",
        "minimum": 40001,
        "maximum": 79999,
        "description": "Platform or domain error code per API_SPEC.md section 15.3."
      },
      "FieldError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "field",
          "message"
        ],
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32",
            "minimum": 40011,
            "maximum": 40099
          }
        }
      },
...
    },
...
  }
}